chore(components): separate tabpanes into different files
:Skip
